The changes made in this commit include: - Modified src/stories/Page.tsx to fix a typo in a comment.
Please summarize the following git diff:
```diff
diff --git a/src/stories/Page.tsx b/src/stories/Page.tsx
index e117483..a1d08ac 100644
--- a/src/stories/Page.tsx
+++ b/src/stories/Page.tsx
@@ -36,7 +36,7 @@ export const Page: React.FC = () => {
         <ul>
           <li>
             Use a higher-level connected component. Storybook helps you compose such data from the
-            "args" of child component stories
+            "args" of child component stories
           </li>
           <li>
             Assemble data in the page component from your services. You can mock these services out
```

In this diff, we can see that there was a typo in a comment in the `src/stories/Page.tsx` file. The word "args" was misspelled as "args". The correction was made to fix the typo and make the comment more accurate.